requirements.txt

flask
torch
transformers
pillow
autocorrect
 flask_cors
 pytesseract
 textblob 